.......... [DEV-681] moved int casts inside ?: operator
git-svn-id: svn://svn.zabbix.com/branches/dev@74112 f4e7fd02-d210-4e34-a1e0-d4093fb66cc2

+++ DEV-681/src/libs/zbxcommon/misc.c
@@ -874,7 +874,7 @@ int	zbx_check_time_period(const char *period, time_t time, int *res)
 	tm = localtime(&time);
 
 	next = strchr(period, ';');
-	while  (SUCCEED == time_period_parse(&tp, period, (int)(NULL == next ? strlen(period) : next - period)))
+	while  (SUCCEED == time_period_parse(&tp, period, (NULL == next ? (int)strlen(period) : (int)(next - period))))
 	{
 		if (SUCCEED == check_time_period(tp, tm))
 			res_total = SUCCEED;	/* no short-circuits, validate all periods before return */
@@ -1698,7 +1698,7 @@ int	zbx_interval_preproc(const char *interval_str, int *simple_interval, zbx_cus
 			new_interval = zbx_malloc(NULL, sizeof(zbx_flexible_interval_t));
 
 			if (SUCCEED != (ret = flexible_interval_parse(new_interval, interval_str,
-					(int)(NULL == delim ? (int)strlen(interval_str) : delim - interval_str))))
+					(NULL == delim ? (int)strlen(interval_str) : (int)(delim - interval_str)))))
 			{
 				zbx_free(new_interval);
 				interval_type = "flexible";
@@ -1716,7 +1716,7 @@ int	zbx_interval_preproc(const char *interval_str, int *simple_interval, zbx_cus
 			memset(new_interval, 0, sizeof(zbx_scheduler_interval_t));
 
 			if (SUCCEED != (ret = scheduler_interval_parse(new_interval, interval_str,
-					(int)(NULL == delim ? (int)strlen(interval_str) : delim - interval_str))))
+					(NULL == delim ? (int)strlen(interval_str) : (int)(delim - interval_str)))))
 			{
 				zbx_free(new_interval);
 				interval_type = "scheduling";
@@ -1733,7 +1733,8 @@ out:
 		if (NULL != error)
 		{
 			*error = zbx_dsprintf(*error, "Invalid %s interval \"%.*s\".", interval_type,
-					(NULL == delim ? (int)strlen(interval_str) : delim - interval_str), interval_str);
+					(NULL == delim ? (int)strlen(interval_str) : (int)(delim - interval_str)),
+					interval_str);
 		}
 
 		flexible_interval_free(flexible);
